requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=2.1.0=pypi_0
argon2-cffi=20.1.0=pypi_0
async-generator=1.10=pypi_0
atomicwrites=1.4.0=pypi_0
attrs=20.3.0=pypi_0
babel=2.9.0=pypi_0
backcall=0.2.0=pypi_0
bleach=3.3.0=pypi_0
ca-certificates=2021.1.19=haa95532_0
certifi=2020.12.5=py38haa95532_0
cffi=1.14.4=pypi_0
chardet=4.0.0=pypi_0
colorama=0.4.4=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
docutils=0.16=pypi_0
entrypoints=0.3=pypi_0
idna=2.10=pypi_0
iniconfig=1.1.1=pypi_0
ipykernel=5.4.3=pypi_0
ipython=7.20.0=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.18.0=pypi_0
jinja2=2.11.3=pypi_0
json5=0.9.5=pypi_0
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.11=pypi_0
jupyter-core=4.7.1=pypi_0
jupyter-server=1.3.0=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-server=2.2.0=pypi_0
lockfile=0.12.2=pypi_0
luigi=3.0.2=pypi_0
markupsafe=1.1.1=pypi_0
mistune=0.8.4=pypi_0
nbclassic=0.2.6=pypi_0
nbclient=0.5.1=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.2=pypi_0
nest-asyncio=1.5.1=pypi_0
notebook=6.2.0=pypi_0
numpy=1.20.1=pypi_0
openssl=1.1.1i=h2bbff1b_0
packaging=20.9=pypi_0
pandas=1.2.1=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.8.1=pypi_0
pickleshare=0.7.5=pypi_0
pip=20.3.3=py38haa95532_0
pluggy=0.13.1=pypi_0
prometheus-client=0.9.0=pypi_0
prompt-toolkit=3.0.14=pypi_0
py=1.10.0=pypi_0
pycparser=2.20=pypi_0
pygments=2.7.4=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.17.3=pypi_0
pytest=6.2.2=pypi_0
python=3.8.5=h5fd99cc_1
python-daemon=2.2.4=pypi_0
python-dateutil=2.8.1=pypi_0
pytz=2021.1=pypi_0
pywin32=300=pypi_0
pywinpty=0.5.7=pypi_0
pyzmq=22.0.2=pypi_0
requests=2.25.1=pypi_0
send2trash=1.5.0=pypi_0
setuptools=52.0.0=py38haa95532_0
six=1.15.0=pypi_0
sniffio=1.2.0=pypi_0
sqlalchemy=1.3.23=pypi_0
sqlite=3.33.0=h2a8f88b_0
terminado=0.9.2=pypi_0
testpath=0.4.4=pypi_0
toml=0.10.2=pypi_0
tornado=6.1=pypi_0
tqdm=4.56.2=pypi_0
traitlets=5.0.5=pypi_0
urllib3=1.26.3=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py38_0
zlib=1.2.11=h62dcd97_4
